+++ borax/calendars/dataset/__init__.py
@@ -11,4 +11,6 @@ _FILE_DICT = {
 
 def get_festival_dataset_path(identifier: str) -> Path:
     """Return the full path for festival dataset csv file."""
+    if identifier not in _FILE_DICT:
+        raise ValueError(f'Festival Dataset {identifier} not found!')
     return Path(__file__).parent / _FILE_DICT.get(identifier)